refactor: remove bot and orchestrator facade indirection
