+++ tests/test_matrix_spaces.py
@@ -456,7 +456,7 @@ async def test_update_config_matrix_space_change_reconciles_without_room_members
 
     with (
         patch.object(Config, "from_yaml", return_value=updated_config),
-        patch("mindroom.orchestrator._identify_entities_to_restart", return_value=set()),
+        patch("mindroom.orchestration.config_updates._identify_entities_to_restart", return_value=set()),
         patch.object(orchestrator, "_setup_rooms_and_memberships", new=AsyncMock()) as mock_setup,
         patch.object(
             orchestrator,
